requirements.txt
backports.csv==1.0.7
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
django-import-export==1.2.0
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.1
psycopg2==2.7.7
pytz==2018.9
PyYAML==5.1
tablib==0.13.0
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
